requirements.txt
streamlit==1.44.1
gspread==5.7.0
google-auth==2.22.0
pandas==2.0.1
